Fix the certificate signature size and clarify cert records size
The size of an Ed25519 signature, as used by the current DNSCrypt
version, is 64 bytes, not 32 bits.
It was not clear to me after the first reading that the 116 bytes
size referred to the size of the signature plus the signed part,
therefore excluding the cert-magic, es-version and
protocol-minor-version. Of course it makes sense, but I think it
can't hurt to write it down.