Complete rewrite of the blocking plugin to make it really fast
This was intended to be nothing but a simple example of a plugin.

However, this example turned out to be useful. And actually used.

So, let's get serious about it and do better than a linear scan
in order to match names and IPs.

A critbit-like data structure is a perfect fit for that kind of list.

So, split blacklists into 4 tries (5-bit variant of qp-tries):
- ips
- name prefixes
- name suffixes
- patterns that can be found anywhere

Linear scans could have been kept for case #4, but this is not a
common use case, and having a single data structure for everything
reduces bloat.

Suffix matching currently works the same way as prefix matching, with
keys and strings being reversed first. However, from a user perspective,
it makes sense to have `*.example.com` match `example.com` and
`www.example.com` but not `abcexample.com`. This will be addressed next.

Meanwhile, the plugin can now seamlessly handle very large blacklists.