FreeTSA
I did not know that the Freetsa server was not on the csv listing since October of last year. I continue to use the server without problems.

Maybe there was a temporary connectivity problem at some point, but it certainly did not go on in time. If you want, you can add it again